+++ tests/devtests/pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+# Put the repo root on sys.path so devtests can ``import hpvsim`` from the
+# working tree and ``from tests.regression.X import Y``, regardless of pytest's
+# working directory.
+pythonpath = ..
+filterwarnings =
+    ignore::pytest.PytestReturnNotNoneWarning
+env =
+    SCIRIS_BACKEND=agg
+markers =
+    slow: marks tests as slow (every test in this folder is slow by definition)
