+++ tests/devtests/test_hiv_coinfection_gate.py
@@ -1,5 +1,13 @@
-import pytest
+"""HIV-HPV coinfection direction gate.
+
+Runs the full Rwanda HIV+HPV anchor sim, which is far too expensive for the
+unit suite. The connector's own arithmetic (CD4 stratification, rel_sus /
+rel_sev multipliers) is unit-tested in ``tests/test_hiv_connector.py``; this
+asserts the emergent population-level direction.
+"""
 import numpy as np
+import pytest
+
 from tests.regression.anchor_hiv_hpv import build_sim
 
 
@@ -13,9 +21,3 @@ def test_hiv_positive_have_higher_hpv_and_cancer():
     assert np.nanmean(res['hpv_prevalence_with_hiv'][sl]) > \
            np.nanmean(res['hpv_prevalence_no_hiv'][sl])
     assert res['cancers_with_hiv'].sum() > 0
-
-
-def test_cd4_high_bins_to_gt200():
-    from hpvsim.hiv import hpv_hiv_connector
-    c = hpv_hiv_connector()
-    assert list(c._cd4_stratum(np.array([594.0, 800.0]))) == [1, 1]
