+++ tests/pytest.ini
@@ -8,4 +8,4 @@ filterwarnings =
 env =
     SCIRIS_BACKEND=agg
 markers =
-    slow: marks tests as slow (skip in unit-run; opt-in with ``-m slow``)
+    slow: reserved; the expensive statistical gates now live in ../devtests
